replace useEffect with util function
